Update the implementation of PAGImageView on the iOS platform.

+++ src/platform/ios/private/PAGDecoderImpl.mm
@@ -73,7 +73,7 @@
 }
 
 - (BOOL)copyFrameTo:(void*)pixels rowBytes:(size_t)rowBytes at:(NSInteger)index {
-  return pagDecoder->readFrame(index, pixels, rowBytes, pag::ColorType::BGRA_8888,
+  return pagDecoder->readFrame((int)index, pixels, rowBytes, pag::ColorType::BGRA_8888,
                                pag::AlphaType::Premultiplied);
 }
 
@@ -86,7 +86,7 @@
   CVPixelBufferLockBaseAddress(pixelBuffer, 0);
   auto rowBytes = CVPixelBufferGetBytesPerRowOfPlane(pixelBuffer, 0);
   auto pixels = CVPixelBufferGetBaseAddress(pixelBuffer);
-  auto success = pagDecoder->readFrame(index, pixels, rowBytes, pag::ColorType::BGRA_8888,
+  auto success = pagDecoder->readFrame((int)index, pixels, rowBytes, pag::ColorType::BGRA_8888,
                                        pag::AlphaType::Premultiplied);
   CVPixelBufferUnlockBaseAddress(pixelBuffer, 0);
   if (!success) {
